Import of previously worked on version of Jupiter iOS app.
This is a pretty rough import but functional (in the sense that the code will build and execute - it doesn’t actually do anything).  Running this version gives an idea of flow.

This version depends on Xcode 5 and uses Xcode 5 storyboards.  Older versions of Xcode will not be able to read the storyboards and will need to be upgraded.